tools/character-editor/pyproject.toml
[project]
name = "character-editor"
version = "0.1.0"
description = "Character editor for nordencult game project"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "pillow>=10.3.0",
    "streamlit-drawable-canvas>=0.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "pages"]
